orders/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pyyaml==6.0.1
jsonschema==4.20.0
django-filter==23.5
